添加手机号相关接口

+++ src/main/java/com/yswl/shop_phone/entity/Phone.java
@@ -5,6 +5,8 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @Setter
 @Getter
@@ -18,7 +20,7 @@ public class Phone {
     private int id;
     private String number;
     private int status;
-    private double money;
+    private BigDecimal money;
     private String operator;
     private String meaning;
     private int beauty;
